backend/requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
sqlalchemy==2.0.49
passlib==1.7.4
bcrypt==4.0.1
PyJWT==2.12.1
pydantic[email]==2.13.2
pydantic-settings==2.13.1
python-dotenv==1.2.2
python-multipart==0.0.26
